pyproject.toml
[tool.poetry]
name = "griptape-flow"
version = "0.10.0"
description = "Python framework for LLM workflows and pipelines."
authors = ["Griptape <hello@griptape.ai>"]
license = "Apache 2.0"
readme = "README.md"
repository = "https://github.com/griptape-ai/griptape-flow"

packages = [
    {include = "griptape"}
]

[tool.poetry.dependencies]
python = "^3.9"
griptape-core = ">= 0.9.1"
python-dotenv = ">=0.21"
openai = ">=0.27"
cohere = ">=4"
attrs = ">=22"
jinja2 = ">=3.1"
jsonschema = ">=4"
marshmallow = ">=3"
marshmallow-enum = ">=1.5"
graphlib = "*"
tiktoken = ">=0.3"
rich = ">=13"
stopit = "*"

[tool.poetry.group.test.dependencies]
griptape-tools = ">= 0.6.0"
pytest = "~=7.1"
pytest-cover = "*"
twine = ">=4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
